separate win/lose logic
